# ddl: speed up the execution time of iterateSnapshotRows to speed up the “add index” operation.