<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Issue Number: close #xxx <!-- REMOVE this line if no issue to close -->

Problem Summary:
speed up iterateSnapshotRows execution time

### What is changed and how it works?

Proposal: [xxx](url) <!-- REMOVE this line if not applicable -->

What's Changed:

How it Works:
Remove `EncodeRecordKey` in iterateSnapshotRows.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
```
// IterInt
create table t(a bigint, b varchar(255), c bigint, primary key(a));
insert into t values(i, i, i);
// IterCommon
create table t(a varchar(255), b varchar(255), c bigint, primary key(a, b));
insert into t values('iabcdefghit', 'dqwertyuiop', i);
```
Do the main code of Benchmark is as follows(insert 1024 rows):
```
	for i := 0; i < b.N; i++ {
		ddl.iterateSnapshotRows(store, 1, tbl, kv.MaxVersion.Ver, minKey, maxKey,
			func(handle kv.Handle, recordKey kv.Key, rawRow []byte) (bool, error) {
				return true, nil
			})
	}
```
Before this PR:
```
BenchmarkIterInt
    1129	   1052079 ns/op	  623986 B/op    10440 allocs/op
    1161	    935688 ns/op	  619816 B/op	   10429 allocs/op
    1244	    870343 ns/op	  610115 B/op	   10416 allocs/op
BenchmarkIterCommon
    1045	   1094081 ns/op	  694312 B/op	   10461 allocs/op
    1023	   1035321 ns/op	  697645 B/op	   10469 allocs/op
    1086	   1141811 ns/op	  687292 B/op	   10453 allocs/op
```
After this PR:
```
BenchmarkIterInt
    1326	    817768 ns/op	  569402 B/op	    8350 allocs/op
    1400	    791739 ns/op	  563227 B/op	    8342 allocs/op
    1324	    826738 ns/op	  569321 B/op	    8350 allocs/op
BenchmarkIterCommon
    1258	    835891 ns/op	  630177 B/op	    8367 allocs/op
    1221	    863048 ns/op	  634258 B/op	    8373 allocs/op
    1233	    887188 ns/op	  633667 B/op	    8372 allocs/op
```
### Release note <!-- bugfixes or new feature need a release note -->

```release-note
Speed up the execution time of iterateSnapshotRows to speed up the “add index” operation.
```